Introduce Scheduler abstract class
