+++ gammapysimulator/scheduler/CTAscheduler.py
@@ -8,10 +8,11 @@
 
 from gammapy.data import Observation, Observations
 from gammapy.irf import load_cta_irfs
+from gammapysimulator.scheduler.scheduler import Scheduler
 
 import numpy as np
 
-class CTAScheduler:
+class CTAScheduler(Scheduler):
     """
     Class that organises Observations and Datasets with CTA IRFs.
     """
@@ -34,38 +35,6 @@ class CTAScheduler:
         self.irfs = load_cta_irfs(self.conf.IRFfilepath)
 
 
-
-    def DefineSchedule(self):
-        """
-        Define the Schedule: start and stop of each Observation (tme bin).
-        """
-
-        # Estimate Number of observations to use numpy linspace
-        ObservationsNumber = (self.conf.timeStop - self.conf.timeStart) / self.conf.timeReso
-        ObservationsNumber = int(np.floor(ObservationsNumber.to('').value))
-
-        # Define starting time of each observation linearly spaced during the night (wrt reference time)
-        ObservationsStart, Livetimes = np.linspace(self.conf.timeStart.value,
-                                                   self.conf.timeStop.value,
-                                                   num = ObservationsNumber,
-                                                   endpoint=False,
-                                                   retstep=True
-                                                   )
-        # Compute Stops and Number
-        ObservationsStop = ObservationsStart+Livetimes
-        ObservationsNumber = ObservationsStart.size
-        
-        # Convert to Astropy Quantity
-        ObservationsStart= ObservationsStart.tolist()* self.conf.timeUnit
-        ObservationsStop = ObservationsStop.tolist() * self.conf.timeUnit
-        Livetimes = Livetimes * self.conf.timeUnit
-        
-        # Log
-        self.log.info(f"Number of Observations: {ObservationsNumber}. Livetimes: {Livetimes}.")
-        self.log.info(f"Observation interval: [{ObservationsStart[0]},{ObservationsStop[-1]}].")
-
-        return ObservationsStart, ObservationsStop
-    
     def SetObservations(self):
         """
         Define the Observations object.
